requirements.txt
argparse
easydict
h5py
matplotlib
numpy
open3d
opencv-python
pyyaml
scipy
tensorboardX
timm==0.4.5
tqdm
transforms3d
termcolor
loguru
einops
comet_ml
calmsize
diffusers
tabulate
ninja

# Dependencies managed by conda
numpy==1.21.2
scipy==1.8.0
pandas==1.4.0
matplotlib==3.5.1
seaborn==0.11.2
scikit-learn==1.0.2
tensorflow-gpu==2.8.0
tensorboard==2.8.0

# Dependencies managed by pip
about-time==3.1.1
absl-py==1.0.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
alive-progress==2.2.0
antlr4-python3-runtime==4.9.3
anyio==3.5.0
astunparse==1.6.3
async-timeout==4.0.2
babel==2.9.1
cachetools==5.0.0
calmsize==0.1.3
ccimport==0.3.7
cftime==1.6.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
comet-ml==3.31.21
commonmark==0.9.1
configobj==5.0.6
crc32c==2.2.post0
cumm-cu111==0.2.8
cupy-cuda111==10.2.0
cycler==0.11.0
cython==0.29.20
dataclasses==0.6
deepspeed==0.6.5
deprecation==2.1.0
diffusers==0.11.1
docker-pycreds==0.4.0
drjit==0.2.1
dulwich==0.20.32
easydict==1.9
einops==0.4.0
everett==3.0.0
fastrlock==0.8
filelock==3.9.0
fire==0.4.0
flatbuffers==2.0
flatten-dict==0.4.2
fonttools==4.29.1
freetype-py==2.3.0
frozenlist==1.3.0
fsspec==2022.2.0
ftfy==6.1.1
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.5.3
gitdb==4.0.9
gitpython==3.1.26
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grapheme==0.6.0
grpcio==1.43.0
h5py==3.6.0
hjson==3.0.2
huggingface-hub==0.11.1
idna==3.3
imageio==2.15.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.10.1
importlib-resources==5.4.0
iopath==0.1.10
jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-packaging==0.12.0
jupyter-server==1.15.6
jupyterlab==3.3.2
jupyterlab-server==2.11.2
keras==2.8.0
keras-preprocessing==1.1.2
kiwisolver==1.3.2
kornia==0.6.6
lark==1.1.2
libclang==14.0.1
llvmlite==0.39.0
loguru==0.6.0
markdown==3.3.6
matplotlib2tikz==0.7.6
meshio==5.3.4
mitsuba==3.0.1
mrcfile==1.3.0
multidict==6.0.2
multipledispatch==0.6.0
mypy-extensions==0.4.3
nbclassic==0.3.7
nbformat==5.2.0
nestargs==0.5.0
netcdf4==1.5.8
networkx==2.6.3
ninja==1.10.2.3
notebook-shim==0.1.0
numba==0.56.0
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
omegaconf==2.2.2
open3d==0.15.2
opencv-python==4.5.5.64
openexr
opt-einsum==3.3.0
pathtools==0.1.2
pccm==0.3.4
plyfile==0.7.4
portalocker==2.5.1
progressbar2==4.0.0
promise==2.3
protobuf==3.19.4
psutil==5.9.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.0
pydeprecate==0.3.1
pyglet==1.5.23
pykeops==1.5
pymcubes==0.1.2
pyopengl==3.1.0
pyparsing==3.0.7
pyquaternion==0.9.9
pyrr==0.10.3
pyrsistent==0.18.1
python-swiftclient==4.0.0
python-utils==3.3.3
pytorch-lightning==1.5.1
pytorch3d==0.3.0
pytz==2021.3
pywavelets==1.2.0
pyyaml==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.3.0
rsa==4.8
ruamel-yaml==0.17.20
ruamel-yaml-clib==0.2.6
semantic-version==2.9.0
sentry-sdk==1.5.4
sharedarray==3.2.1
shortuuid==1.0.8
simple-parsing==0.0.18
simplejson==3.18.0
smmap==5.0.0
sniffio==1.2.0
tabulate==0.8.9
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardx==2.4.1
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tflearn==0.5.0
tfrecord==1.14.1
threadpoolctl==3.1.0
tifffile==2022.2.2
tikzplotlib==0.10.1
tomlkit==0.10.0
torchmetrics==0.7.2
tqdm==4.62.3
trimesh==3.10.1
typing-extensions==4.2.0
typing-inspect==0.7.1
urllib3==1.26.8
wandb==0.12.10
webcolors==1.11.1
websocket-client==1.2.3
werkzeug==2.0.3
wrapt==1.13.3
wurlitzer==3.0.2
yacs==0.1.8
yarl==1.7.2
yaspin==2.1.0
